behavioral: update DataLoaders and resolvers for session/history split
